requirements.txt
Flask==2.0.2
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
pytangle==0.0.2
python_dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests_oauthlib==1.3.1
seaborn==0.12.0
tweepy==4.10.1
